+++ web/static/js/brando/brando/components/autoslug.js
@@ -8,7 +8,12 @@ class Autoslug {
         $('[data-slug-from]').each((index, elem) => {
             var slugFrom = $(elem).attr('data-slug-from');
             $('[name="' + slugFrom + '"]').slugIt({
-                output: $(elem)
+                output: $(elem),
+                map: { 'æ': 'ae', 'ø': 'oe', 'å': 'aa' },
+                space: '-',
+                after: function(slug) {
+                    return slug.replace(/'/g, '');
+                }
             });
         });
     }